backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
torch==2.1.2
Pillow==10.2.0
git+https://github.com/openai/CLIP.git
python-multipart==0.0.6
python-dotenv==1.0.1
numpy==1.26.4